requirements.txt
pillow
pathlib